refactor: DispatcherServlet만 사용하도록 SpringServletContainerInitializer 제거
