--- mvc/src/main/java/web/org/springframework/web/SpringServletContainerInitializer.java
@@ -1,41 +1,0 @@
-package web.org.springframework.web;
-
-import core.org.springframework.util.ReflectionUtils;
-import jakarta.servlet.ServletContainerInitializer;
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.HandlesTypes;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-@HandlesTypes(WebApplicationInitializer.class)
-public class SpringServletContainerInitializer implements ServletContainerInitializer {
-
-    @Override
-    public void onStartup(Set<Class<?>> webAppInitializerClasses, ServletContext servletContext)
-            throws ServletException {
-        final List<WebApplicationInitializer> initializers = new ArrayList<>();
-
-        if (webAppInitializerClasses != null) {
-            for (Class<?> waiClass : webAppInitializerClasses) {
-                try {
-                    initializers.add((WebApplicationInitializer)
-                            ReflectionUtils.accessibleConstructor(waiClass).newInstance());
-                } catch (Throwable ex) {
-                    throw new ServletException("Failed to instantiate WebApplicationInitializer class", ex);
-                }
-            }
-        }
-
-        if (initializers.isEmpty()) {
-            servletContext.log("No Spring WebApplicationInitializer types detected on classpath");
-            return;
-        }
-
-        for (WebApplicationInitializer initializer : initializers) {
-            initializer.onStartup(servletContext);
-        }
-    }
-}
