--- mvc/src/main/java/web/org/springframework/web/WebApplicationInitializer.java
@@ -1,8 +1,0 @@
-package web.org.springframework.web;
-
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletException;
-
-public interface WebApplicationInitializer {
-    void onStartup(ServletContext servletContext) throws ServletException;
-}
