+++ mvc/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,1 @@
-web.org.springframework.web.SpringServletContainerInitializer
+web.org.springframework.web.DispatcherServletInitializer
